ServiceQueueStatistics implemented

+++ iqueue/src/main/java/pt/ipl/isel/ps/iqueue/repository/ServiceQueueStatisticRowMapper.java
@@ -13,7 +13,7 @@ public class ServiceQueueStatisticRowMapper implements RowMapper<ServiceQueueSta
         return new ServiceQueueStatistic(rs.getInt("attendanceCount"),
                 rs.getInt("averageWaitingSeconds"),
                 rs.getInt("averageAttendanceSeconds"),
-                rs.getInt("averageRate"),
+                rs.getFloat("averageRate"),
                 rs.getInt("quitCount"));
     }
 }
